requirements.txt
Django~=5.0.2
djangorestframework~=3.14.0
psycopg2-binary~=2.9.9
pytest~=8.0.0
pytest-django~=4.8.0
pandas~=2.2.0